working on printDI function.

+++ chermes/inc/structures.h
@@ -10,11 +10,11 @@ struct clusterInfo {
 
 // Represents the data for a single signal.
 struct signalData {
-    int8_t signalType;    // Type of the signal (TDC=1,Pixel=2,GTS=3)
-    uint8_t xPixel;       // X-coordinate of the pixel
-    uint8_t yPixel;       // Y-coordinate of the pixel
-    double ToaFinal;      // Time of Arrival (final value)
-    double TotFinal;      // Time over Threshold (final value)
+    uint8_t signalType; // Type of the signal (TDC=1,Pixel=2,GTS=3)
+    uint8_t xPixel;     // X-coordinate of the pixel
+    uint8_t yPixel;     // Y-coordinate of the pixel
+    double ToaFinal;    // Time of Arrival in seconds (final value)
+    float TotFinal;     // Time over Threshold in nano seconds (final value)
 };
 
 #endif
